requirements.txt
requests
httpx
pandas
python-dotenv